refactor: Move controllers into cluster directory
Intermediate commit to let git keep track of all files
