+++ pkg/clients/cluster/argocd.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package clients
+package cluster
 
 import (
 	"context"
