+++ pkg/clients/cluster/mock/applications/mock.go
