+++ pkg/clients/cluster/mock/applicationsets/mock.go
