+++ pkg/clients/cluster/mock/generate.go
