+++ pkg/controller/cluster/applications/comp.go
@@ -10,7 +10,7 @@ import (
 	"github.com/google/go-cmp/cmp/cmpopts"
 
 	"github.com/crossplane-contrib/provider-argocd/apis/cluster/applications/v1alpha1"
-	"github.com/crossplane-contrib/provider-argocd/pkg/clients/applications"
+	"github.com/crossplane-contrib/provider-argocd/pkg/clients/cluster/applications"
 )
 
 // IsApplicationUpToDate converts ApplicationParameters to its ArgoCD Counterpart and returns if they equal
