+++ pkg/controller/cluster/applications/comp_test.go
