+++ pkg/controller/cluster/cluster/controller.go
@@ -42,8 +42,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/crossplane-contrib/provider-argocd/apis/cluster/cluster/v1alpha1"
-	"github.com/crossplane-contrib/provider-argocd/pkg/clients"
-	"github.com/crossplane-contrib/provider-argocd/pkg/clients/cluster"
+	clients "github.com/crossplane-contrib/provider-argocd/pkg/clients/cluster"
+	"github.com/crossplane-contrib/provider-argocd/pkg/clients/cluster/cluster"
 	"github.com/crossplane-contrib/provider-argocd/pkg/features"
 )
